requirements.txt
scipy
requests
orjson
tqdm==4.66.5
pymongo
motor
ollama
numpy
chromadb
scikit-learn
pydriller
pyyaml
sentence_transformers
einops
pandas
accelerate
hf_xet